% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acr.R
\name{acrshort}
\alias{acrshort}
\alias{acrfull}
\alias{acrlong}
\title{FUNCTION_TITLE}
\usage{
acrshort(term)

acrfull(term)

acrlong(term)
}
\arguments{
\item{term}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
