% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/verbs.R
\name{add_acronym}
\alias{add_acronym}
\title{FUNCTION_TITLE}
\usage{
add_acronym(entry, label, long)
}
\arguments{
\item{entry}{PARAM_DESCRIPTION}

\item{label}{PARAM_DESCRIPTION}

\item{long}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
