% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/verbs.R
\name{add_definition}
\alias{add_definition}
\title{FUNCTION_TITLE}
\usage{
add_definition(entry, name, desc)
}
\arguments{
\item{entry}{PARAM_DESCRIPTION}

\item{name}{PARAM_DESCRIPTION}

\item{desc}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
