% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glossary_utils.R
\name{glossary_to_table}
\alias{glossary_to_table}
\title{FUNCTION_TITLE}
\usage{
glossary_to_table(entries = get_entries())
}
\arguments{
\item{entries}{PARAM_DESCRIPTION, Default: get_entries()}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
