% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glossary_utils.R, R/io.R
\name{load_entries}
\alias{load_entries}
\alias{new_glossary}
\alias{export_glossary}
\title{Input/Output}
\usage{
load_entries(file = "", ...)

new_glossary(file = "", ...)

export_glossary(file, ...)
}
\arguments{
\item{file}{character, path to file storing glossary terms. Default: ''}

\item{...}{additional arguments passed to \link[yaml:read_yaml]{read_yaml}}
}
\value{
glossary
}
\description{
Utility functions reading/exporting the glossary
}
\details{
new_glossary returns a new glossary object that is separate from
the internal glossary object.
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
Other io: 
\code{\link{glossary_to_table}()}
}
\concept{io}
