% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gls.R
\name{gls}
\alias{gls}
\alias{glspl}
\alias{Gls}
\alias{Glspl}
\title{Inline Verbs}
\usage{
gls(term)

glspl(term)

Gls(term)

Glspl(term)
}
\arguments{
\item{term}{character, name of the term.}
}
\value{
character
}
\description{
Inline verbs to create
glossary outputs from loaded glossary object.
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
Other inline: 
\code{\link{acrshort}()}
}
\concept{inline}
