% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/verbs.R
\name{get_entries}
\alias{get_entries}
\alias{get_entry}
\alias{entry_names}
\title{Query Glossary}
\usage{
get_entries(type = c("definition", "acronym"))

get_entry(entry)

entry_names()
}
\arguments{
\item{type}{character, term type 'definition' or 'acronym'}

\item{entry}{character, entry name}
}
\value{
\code{get_entry()} returns a list with contents of a single entry

\code{get_entries()} returns a list of entries by type

\code{entry_name()} returns a character vector of the stored entries
}
\description{
Query glossary entries
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\concept{query}
