% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/verbs.R
\name{used_entries}
\alias{used_entries}
\title{Query Used Terms in Document}
\usage{
used_entries(type = c("definition", "acronym"))
}
\arguments{
\item{type}{character, term type 'definition' or 'acronym'}
}
\value{
returns a list of entries by type used in the document
}
\description{
Query used glossary entries
}
