% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/verbs.R
\name{add_definition}
\alias{add_definition}
\alias{add_acronym}
\alias{update_entry}
\alias{remove_entry}
\title{Glossary actions}
\usage{
add_definition(entry, name, desc)

add_acronym(entry, name, desc)

update_entry(entry, name, desc)

remove_entry(entry)
}
\arguments{
\item{entry}{character, entry name}

\item{name}{character, short name}

\item{desc}{character, long description}
}
\description{
Add, remove,  update or retrieve terms in the glossary.
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\concept{verbs}
