% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acr.R
\name{acrshort}
\alias{acrshort}
\alias{acrfull}
\alias{acrlong}
\title{Inline Acronym Verbs}
\usage{
acrshort(term)

acrfull(term)

acrlong(term)
}
\arguments{
\item{term}{character, name of the term.}
}
\value{
character
}
\description{
Verbs to invoke inline outputs for acronyms
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
Other inline: 
\code{\link{gls}()}
}
\concept{inline}
