% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glossary_utils.R, R/io.R
\name{load_entries}
\alias{load_entries}
\alias{use_glossary}
\alias{glossary_to_yaml}
\title{FUNCTION_TITLE}
\usage{
load_entries(file = "", ...)

use_glossary(file = "", ...)

glossary_to_yaml(file, ...)
}
\arguments{
\item{file}{PARAM_DESCRIPTION, Default: ''}

\item{...}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
