% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gls.R
\name{gls}
\alias{gls}
\alias{Gls}
\alias{glspl}
\alias{Glspl}
\title{FUNCTION_TITLE}
\usage{
gls(term)

Gls(term)

glspl(term)

Glspl(term)
}
\arguments{
\item{term}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
