% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/verbs.R
\name{get_entry}
\alias{get_entry}
\alias{remove_entry}
\alias{update_entry}
\alias{get_entries}
\alias{entry_names}
\title{FUNCTION_TITLE}
\usage{
get_entry(entry)

remove_entry(entry)

update_entry(entry, name, desc)

get_entries(type = c("definition", "acronym"))

entry_names()
}
\arguments{
\item{entry}{PARAM_DESCRIPTION}

\item{name}{PARAM_DESCRIPTION}

\item{desc}{PARAM_DESCRIPTION}

\item{type}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
